1. Enhance window_area_capture source    - Sync window area capture property changes    - Replace QPixmap-based area selection with QWidget overlay
2. Update media handling
   - Add SOOP media source refresh
   - Define video/audio channels for breaktime

3. Cleanup and fix bugs
   - Fix properties height issue
   - Fix cursor style on popup close button

4. Add scene backup logic

5. Sync obs-studio changes
  - obs-studio : Update audio monitoring type handling
  - obs-browser : Improve browser handling

+++ src/Blocks/CDockTitle.cpp
@@ -262,8 +262,8 @@ void AFDockTitle::Initialize(bool onlyPopup, QString text, int BlockType, bool n
     setWindowTitle(text);
 }
 
-void AFDockTitle::InitializeCustom(QString customName, QString customUuid, bool minmax, bool threeDots, bool closeButton)
-{    
+void AFDockTitle::InitializeCustom(QString customName, QString customUuid, bool minmax, bool threeDots, bool closeButton, bool needQuestionMark, const QString& questionMarkToolTip) 
+{
     setProperty("customUuid", customUuid);
     m_blockType = -1;
 
@@ -277,8 +277,11 @@ void AFDockTitle::InitializeCustom(QString customName, QString customUuid, bool 
     
     if (closeButton)
         ui->pushButton_Close->show();
-    DeleteQuestionMarkButton();
-    DeleteTitleIcon();
+
+    if (!needQuestionMark)
+        DeleteQuestionMarkButton();
+    else
+        ui->pushButton_QuestionMark->SetExplanationText(questionMarkToolTip, AFQInfoTooltipButton::ToolTipPos::BottomLeft);
 
     setWindowTitle(customName);
     MinMaxButton(minmax);
@@ -332,7 +335,7 @@ void AFDockTitle::UpdateTitleLabel()
 
     int finalWidth = qMin(titleWidth, adjustTitleWidth);
 
-    if (finalWidth < 140)
+    if ((finalWidth < 140) && !ui->pushButton_QuestionMark)
         finalWidth = 140;
 
     if (ui->label->width() != finalWidth)
